feat: moved some files + began readme doc
