+++ src/init/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod login;
 pub(crate) mod session;
+pub mod singletons;
 mod sync;
 pub(crate) mod workers;
