+++ src/models/mod.rs
@@ -1,5 +1,5 @@
 pub(crate) mod async_requests;
 pub mod device;
 pub mod event_bridge;
-pub mod requests;
+pub mod events;
 pub mod state_updater;
