+++ src/room/mod.rs
@@ -2,6 +2,7 @@ pub(crate) mod frontend_events;
 pub(crate) mod invited_room;
 pub(crate) mod joined_room;
 pub(crate) mod message;
+pub(crate) mod notifications;
 pub(crate) mod room_filter;
 pub(crate) mod room_screen;
 pub(crate) mod rooms_list;
